refactor: create utils module
